Attempt at making the help system play along
We just make sure that the pager does not try to page. We get
all of the manual back, which is of course suboptimal at best.

In the future we should be either linking to statically served help,
or render smaller bits of the manual, for example in help bubbles.