Added modified HELP_SHOW_MATCHES
This version uses StripEscapeSequences when displaying a hyperlink to the unique match, and displays the full list of matches in case of multiple matches.

We assume that the Pages is `tail', so in case of multiple matches, one should specify that it should display from line 2, to avoid displaying 10 last entries by default. Since `tail' does not work from line 1, an empty line has been added in the beginning of the outpuyt.

This will fix #12.